refactor: moved getLiveEditorRedirectionUrl to a separate file
